requirements.txt
altair
argon2-cffi==21.1.0
astor==0.8.1
attrs==21.2.0
backcall==0.2.0
backports.zoneinfo==0.2.1
base58==2.1.1
bleach==4.1.0
blinker==1.4
blis
cachetools==4.2.4
catalogue==2.0.6
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
click==7.1.2
cycler==0.11.0
cymem==2.0.6
debugpy==1.5.1
decorator==5.1.0
defusedxml==0.7.1
entrypoints==0.3
filelock==3.3.2
funcy==1.16
future==0.18.2
gensim
gitdb==4.0.9
GitPython==3.1.24
huggingface-hub==0.1.1
idna==3.3
importlib-resources==5.4.0
ipykernel==6.5.0
ipython==7.29.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
jedi==0.18.0
Jinja2==3.0.2
joblib==1.1.0
jsonschema==4.2.1
jupyter-client==7.0.6
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
kiwisolver==1.3.2
langcodes==3.2.1
MarkupSafe==2.0.1
matplotlib
matplotlib-inline==0.1.3
mistune==0.8.4
murmurhash==1.0.6
mysql
nbclient==0.5.5
nbconvert==6.2.0
nbformat==5.1.3
nest-asyncio==1.5.1
nltk==3.6.5
notebook==6.4.5
numexpr
numpy==1.19.5
packaging==21.2
pandas
pandocfilters==1.5.0
parso==0.8.2
pathy==0.6.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
plotly==5.3.1
preshed==3.0.6
prometheus-client==0.12.0
prompt-toolkit==3.0.22
protobuf==3.19.1
ptyprocess==0.7.0
pyarrow
pycparser==2.21
pydantic==1.8.2
pydeck
Pygments==2.10.0
pyLDAvis
pyparsing==2.4.7
pyrsistent==0.18.0
python-dateutil==2.8.2
python-dotenv==0.19.1
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
pyzmq==22.3.0
regex==2021.11.2
requests==2.26.0
sacremoses==0.0.46
scikit-learn==1.0.1
scikit-learn-extra==0.2.0
scipy==1.7.2
seaborn==0.11.2
Send2Trash==1.8.0
sentence-transformers==2.1.0
sentencepiece==0.1.96
six==1.16.0
sklearn==0.0
smart-open==5.2.1
smmap==5.0.0
spacy==3.2.0
spacy-legacy==3.0.8
spacy-loggers==1.0.1
srsly==2.4.2
streamlit==1.1.0
streamlit-tags==1.2.8
streamlit-yellowbrick==0.0.3
tenacity==8.0.1
terminado==0.12.1
testpath==0.5.0
thinc==8.0.13
threadpoolctl==3.0.0
tokenizers==0.10.3
toml==0.10.2
toolz==0.11.2
torch==1.10.0
torchvision==0.11.1
tornado==6.1
tqdm==4.62.3
traitlets==5.1.1
transformers==4.12.3
typer==0.4.0
typing-extensions==3.10.0.2
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.7
validators==0.18.2
wasabi==0.8.2
watchdog==2.1.6
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.2
yellowbrick==1.3.post1
zipp==3.6.0
